backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dateutil==2.8.2
aiofiles==23.2.1
requests==2.31.0
boardgamegeek2==0.15.0
